taskid int -> long GUID.isGloabal true -> false

+++ DataTransfer/src/edu/fudan/se/crowdservice/wrapper/ResponseWrapper.java
@@ -9,10 +9,10 @@ import java.util.ArrayList;
  * Created by Jiahuan on 2015/1/22.
  */
 public class ResponseWrapper implements Serializable {
-    public final int taskId;
+    public final long taskId;
     public final ArrayList<KeyValueHolder> keyValueHolders;
 
-    public ResponseWrapper(int taskId, ArrayList<KeyValueHolder> keyValueHolders) {
+    public ResponseWrapper(long taskId, ArrayList<KeyValueHolder> keyValueHolders) {
         this.taskId = taskId;
         this.keyValueHolders = keyValueHolders;
     }
